pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "ibisml"
description = "Tools for developing ML pipelines using Ibis"
readme = "README.rst"
requires-python = ">= 3.8"
license = {text = "Apache 2.0"}
dynamic = ["version"]
dependencies = [
    "ibis-framework"
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401", "F403"]
